requirements.txt
alembic==1.7.7
click==8.1.3
Flask==1.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
greenlet==1.1.2
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.0.1
Jinja2==2.10.1
Mako==1.2.0
MarkupSafe==2.0.1
psycopg2==2.9.3
pydantic==1.9.0
PyJWT==2.4.0
six==1.16.0
SQLAlchemy==1.4.36
typing-extensions==4.2.0
Werkzeug==0.16.1
zipp==3.8.0
